requirements.txt
torch==2.0.1
torchvision==0.15.2
timm==0.9.2
numpy==1.25.2
pillow==10.0.0
opencv-python==4.8.0.76
scipy==1.11.1
omegaconf==2.3.0
einops==0.6.1
transformers==4.31.0
